+++ Changes.md
@@ -7,9 +7,11 @@
 - FIXES: Issues when passing a block to chaining scope.
 - FIXES: Enums properly handle default values.
 - FIXES: Enums macro methods now are dirty tracked.
+- FIXES: #17: overloads inspect to show objects in a has_many scope.
 - `reload!` and session methods support client request options.
 - Proactively deletes possible matching cache keys on non GET requests.
 - Parse::File now has a `force_ssl` option that makes sure all urls returned are `https`.
+- Documentation
 
 ### 1.5.1
 - BREAKING CHANGE: The default `has_many` implementation is `:query` instead of `:array`.